refactoring craftassist specific methods from base_util -> craftassist_shared_utils
