+++ droidlet/interpreter/craftassist/spatial_reasoning.py
@@ -6,7 +6,8 @@ import numpy as np
 
 import droidlet.base_util
 from droidlet.perception.craftassist import rotation, heuristic_perception
-from droidlet.base_util import to_block_center, to_block_pos, arrange
+from droidlet.base_util import to_block_center, to_block_pos
+from droidlet.shared_data_struct.craftassist_shared_utils import arrange
 from droidlet.shared_data_structs import ErrorWithResponse
 
 DEFAULT_NUM_STEPS = 5
